+++ tests/extras/scheme/class_tests.scm
@@ -82,7 +82,7 @@
 ;; 5. bind from Scheme
 (define (adder a b) (+ a b))
 (define add10 (adder.bind #f 10))
-(test "bind procedure from Scheme" 25 (add10 15))
+(test "bind procedure from Scheme" 25.0 (add10 15))
 
 (define-class <Greeter>
   Greeter