+++ tests/tests.js
@@ -4,6 +4,7 @@ import { runInteropTests } from './functional/interop_tests.js';
 import { runQuasiquoteTests } from './functional/quasiquote_tests.js';
 import { runDefineTests } from './functional/define_tests.js';
 import { runJsGlobalTests } from './functional/js_global_tests.js';
+import { runInteropConversionTests } from './functional/interop_conversion_tests.js';
 
 export { runUnitTests };
 
@@ -18,4 +19,5 @@ export function runAllTests(interpreter, logger) {
   runQuasiquoteTests(interpreter, logger);
   runDefineTests(interpreter, logger);
   runJsGlobalTests(interpreter, logger);
+  runInteropConversionTests(interpreter, logger);
 }